+++ Utilities/CRCCheck.cpp
@@ -1,16 +1,16 @@
 #include "CRCCheck.hpp"
 #include "../RMTMessageDefs.hpp"
 
-uint32_t crc4_itu(uint32_t data)
+uint16_t crc4_itu(uint16_t data)
 {
     for (int i = 0; i < IR::detail::Msg_content_bits; ++i)
         data = (data >> 1) ^ ((-(data & 1)) & 0xC);
     return data;
 }
 
-uint32_t crc8_maxim(const uint32_t *data, const uint32_t length)
+uint16_t crc8_maxim(const uint16_t *data, const uint16_t length)
 {
-    uint32_t i, length1 = length, crc = 0;
+    uint16_t i, length1 = length, crc = 0;
     while (length1--)
     {
         crc ^= *data++; // crc ^= *data; data++;
@@ -20,9 +20,9 @@ uint32_t crc8_maxim(const uint32_t *data, const uint32_t length)
     return crc;
 }
 
-uint32_t crc12_cdma(const uint32_t *data, const uint32_t length)
+uint16_t crc12_cdma(const uint16_t *data, const uint16_t length)
 {
-    uint32_t i, length1 = length, crc = 0;
+    uint16_t i, length1 = length, crc = 0;
     while (length1--)
     {
         crc ^= *data++; // crc ^= *data; data++;
